waterquality_project/requirements.txt
Django==5.2.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
openpyxl==3.1.2
numpy==1.24.3
scipy==1.11.1 